+++ Visual Studio Projects/ZWaveJS.NET/Demo Application/Information.xaml.cs
@@ -1,0 +1,10 @@
+namespace Demo_Application;
+
+public partial class Information : ContentPage
+{
+	public Information()
+	{
+		InitializeComponent();
+	
+	}
+}